More more tests

+++ src/main/java/com/nbd/plantstore/services/AddressService.java
@@ -3,6 +3,7 @@ package com.nbd.plantstore.services;
 
 import com.nbd.plantstore.Exceptions.addressNotExistById;
 import com.nbd.plantstore.Exceptions.addressNotExistbyAll;
+import com.nbd.plantstore.Exceptions.canNotDeleteAddress;
 import com.nbd.plantstore.Exceptions.thisAddressAlreadyExist;
 import com.nbd.plantstore.entities.Address;
 import com.nbd.plantstore.repositories.AddressRepository;
@@ -17,12 +18,17 @@ public class AddressService {
     @Autowired
     private AddressRepository addressRepository;
 
+    @Autowired
+    private ClientService clientService;
+
 
     public Address findAddressByIdIfExist (Long id) {
         return addressRepository.findById(id).orElseThrow(() -> new addressNotExistById(id));
     }
 
     public void deleteAddress(Long id) {
+        if(clientService.checkIsClientExistByAddress(id))
+            throw new canNotDeleteAddress();
         addressRepository.delete(findAddressByIdIfExist(id));
     }
 
@@ -32,16 +38,22 @@ public class AddressService {
 
     @Transactional
     public void streetUpdate (String newStreet, Address address) {
+        if(addressRepository.existsAddressByCityAndStreetAndStreet_number(address.getCity(), address.getStreet(), address.getStreet_number()))
+            throw new thisAddressAlreadyExist(address.getCity(), address.getStreet(), address.getStreet_number());
         addressRepository.updateStreet(newStreet, address.getId());
     }
 
     @Transactional
     public void cityUpdate (String newCity, Address address) {
+        if(addressRepository.existsAddressByCityAndStreetAndStreet_number(address.getCity(), address.getStreet(), address.getStreet_number()))
+            throw new thisAddressAlreadyExist(address.getCity(), address.getStreet(), address.getStreet_number());
         addressRepository.updateStreet(newCity, address.getId());
     }
 
     @Transactional
     public void streetNumUpdate (Integer newStreetNum, Address address) {
+        if(addressRepository.existsAddressByCityAndStreetAndStreet_number(address.getCity(), address.getStreet(), address.getStreet_number()))
+            throw new thisAddressAlreadyExist(address.getCity(), address.getStreet(), address.getStreet_number());
         addressRepository.updateStreetNum(newStreetNum, address.getId());
     }
